MOVEONLY: Move to consensus/consensus.h:
-Most consensus constants that are not in script already
-Functions declaration for most of what's necessary for verifying transactions, block headers and blocks
